% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutomaticMapping.R
\name{AutomaticMapping}
\alias{AutomaticMapping}
\title{AutomaticMapping}
\usage{
AutomaticMapping(x,mapping=NULL,from=NULL,to=NULL)
}
\arguments{
\item{x}{MAgPIE object}

\item{mapping}{A array or data.frame containing a mapping query or mapping name}

\item{from}{Only required if query is not NULL. Column of the query with
original dimnames of the incoming dataset}

\item{to}{Only required if query is not NULL. Column of the query with the
target dimnames of the outcoming dataset}
}
\description{
Function automatically finds fitting mapping for provided MAgPIE object for a given target aggregation.
}
\author{
Benjamin Leon Bodirsky
}
