% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_per_region.R
\name{cluster_per_region}
\alias{cluster_per_region}
\title{Cluster per Region}
\usage{
cluster_per_region(cdata, ncluster, weight = NULL)
}
\arguments{
\item{cdata}{a cluster data file as produced by cluster_base}

\item{ncluster}{The desired total number of clusters.}

\item{weight}{named vector with weighting factors for each region for the cluster distribution 
,e.g. weight=c(AFR=3,EUR=0.5). weight > 1 will grant more cluster to a region and
weight < 1 less cluster than by default.}
}
\value{
A matrix with regions in rows and number of cells and clusters in
columns
}
\description{
This function calculates an appropriate number of clusters per region as it
is needed for mag_kmeans
}
\seealso{
\code{\link{mag_kmeans}}, \code{\link{cluster_base}},
\code{\link{clusterspam}}
}
\author{
Jan Philipp Dietrich
}
