% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterspam.R
\name{clusterspam}
\alias{clusterspam}
\title{Main MAgPIE clustering function}
\usage{
clusterspam(
  lr,
  hr = "0.5",
  ifolder = ".",
  ofolder = ".",
  cfiles = c("lpj_yields", "lpj_airrig", "transport_distance"),
  years2use = "y1995",
  spatial_header = NULL,
  use_cache = TRUE,
  weight = NULL,
  seed = 42
)
}
\arguments{
\item{lr}{low resolution containing the clustering mode as first letter in
the name, e.g. h100. Available modes are at the moment n: normed kmeans
clustering, h: hierarchical complete linkage clustering, s: hierarchical
single linkage clustering, w: hierarchical Ward clustering and c: combined
hierarchical/k-means-clustering. In the latter hierarchical complete-linkage
clustering is used to determine the distribution of clusters among regions
and normed-k-means clustering is used to determine clusters within regions}

\item{hr}{high resolution, e.g. 0.5}

\item{ifolder}{input folder where the MAgPIE input files are located}

\item{ofolder}{output folder where spam relation matrix should be written
to.}

\item{cfiles}{a vector containing the names of the MAgPIE input files
(beginning of the name is enough)}

\item{years2use}{A vector with years which should be taken into account for
the clustering}

\item{spatial_header}{A vector of the form c("REG.1","REG.2") (region name,
cell number) with entries for each spatial entity of the MAgPIE input files
which should be used to replace the names given in the inputs (required for
flexible region aggregation as here region names might change.). If set to
NULL the original information is used.}

\item{use_cache}{Read data from cache file if available (dangerous as changes
in settings will not be considered if an existing cache file is found).}

\item{weight}{named vector with weighting factors for each region for the cluster distribution 
,e.g. weight=c(AFR=3,EUR=0.5). weight > 1 will grant more cluster to a region and
weight < 1 less cluster than by default.}

\item{seed}{a single value, interpreted as an integer, or NULL, to define seed for random calculations}
}
\value{
A spam relation matrix
}
\description{
This is the main MAgPIE clustering function which you should use if you want
to create a spam relation matrix based on clustering methods
}
\details{
The Spam relation matrix is written at the same time in the output folder.
}
\seealso{
\code{\link{cluster_per_region}}, \code{\link{mag_kmeans}},
\code{\link{mag_hierarchical}}
}
\author{
Jan Philipp Dietrich
}
