% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create Grid}
\usage{
create_grid(res_in,res_out,fname=NULL,folder=NULL)
}
\arguments{
\item{res_in}{input resolution in degree (e.g. 0.5)}

\item{res_out}{output resolution in degree (e.g. 12)}

\item{fname}{file name of a file the output should be written to. If
fname=NULL the relation matrix is returned by the function, if
fname="default" the default name format res_in-to-res_out_sum.spam is used}

\item{folder}{folder the file should be written to}
}
\value{
If fname=NULL, the relation matrix, otherwise nothing is returned.
}
\description{
Create a spam relation matrix between two regular grid resolutions (measured
in degree)
}
\examples{

 \dontrun{rel <- create_grid(0.5,12.0)}

}
\seealso{
\code{\link{create_spam}}, \code{\link{read.spam}},
\code{\link{write.spam}}
}
\author{
Jan Philipp Dietrich
}
