% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spam.R
\name{create_spam}
\alias{create_spam}
\title{Create SPAM}
\usage{
create_spam(x,rel,year=NULL,elem=NULL,fname=NULL)
}
\arguments{
\item{x}{Vector, MAgPIE object or file name of a MAgPIE object. If MAgPIE
object it must either contain only a single year and a single data column or
year and elem have to be defined.}

\item{rel}{Input spam relation matrix or file name of a input spam relation
matrix}

\item{year}{Year which should be used for calculation (has to be specified
if the MAgPIE object contains more than 1 year)}

\item{elem}{Data column which should be used for calculation (has to be
specified if the MAgPIE object contains more than 1 data column)}

\item{fname}{file name of a file the output should be written to. If
fname=NULL the spam matrix is returned by the function}
}
\value{
If fname=NULL, the spam matrix, otherwise nothing is returned.
}
\description{
Creates a spam relation matrix based on another relation matrix and a 1D
MAgPIE object (only 1 year and 1 data column)
}
\examples{

 \dontrun{rel2 <- create_spam(x,rel)}

}
\seealso{
\code{\link{speed_aggregate}}, \code{\link{read.spam}},
\code{\link{write.spam}}
}
\author{
Jan Philipp Dietrich
}
