% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_intensity.R
\name{downscale_intensity}
\alias{downscale_intensity}
\title{downscale intensity}
\usage{
downscale_intensity(nomiR, denomR, nomibaseC, denomC, mapping,
convmethod="exp", convrate="standard", replacemissing=FALSE,
adjustnomiR=TRUE, adjustdenomR=TRUE, adjustshare="by nomiC")
}
\arguments{
\item{nomiR}{the data in the nominator (e.g. GDP,Energy,Emissions) for all
regions for all relevant years.}

\item{denomR}{the data in the denominator (e.g. Population,GDP,Energy) for
all regions for all relevant years.}

\item{nomibaseC}{base year data of the nominator at country (or cell, if
applied as country to cell) level, e.g. GDP in 2010}

\item{denomC}{data for the denominator at country level for all relevant
years, e.g. population, GDP, Energy}

\item{mapping}{file for mapping countries to regions and vice versa.}

\item{convmethod}{the method of convergence of the intensities. Exponential
("exp") or linear ("lin").}

\item{convrate}{A number between 0 and 1 which specifies the degree of
convergence in the last data year. 1 means total convergence. Standard is
1-exp(-1) for exponential (i.e. the difference has decayed to e^(-1) of its
initial value) and 1 for linear.}

\item{replacemissing}{specifies whether unavailable data should be replaced
with region averages or not. If TRUE, downscale_intensity uses all available
information of the input data, i.e. it uses data from nomibaseC and denomC
even if the corresponding value of denomC resp. nomibaseC is not available
by using region averages. If FALSE, information is ignored when it is not
available for one of the two.Results will differ slightly depending on the
choice of replacemissing. This is because of the different information
content used.}

\item{adjustnomiR}{if replacemissing is FALSE, setting adjustnomiR to TRUE
will adjust the input data by substracting the data of those countries that
are ignored from their respective regions.}

\item{adjustdenomR}{same as adjustnomiR for the denominator.}

\item{adjustshare}{specifies how discrepancies between the downscaled data
and the regional Remind output are attributed to the countries. "by nomiC"
uses the denominator as weight (e.g. countries with higer GDP receive a
bigger share of the difference term). "by growth" uses the growth in every
timestep as weight, e.g. countries that have higher growth receive a bigger
share of the difference term.}
}
\value{
\item{downscaled_intensity}{Magpie-object containing the downscaled
intensity for all available years and countries.}
}
\description{
Downscaling of intensities like GDP/cap, Energy/GDP, Emissions/Energy from
regional to country level. Intensities have the structure
dataNominator/dataDenominator.
}
\details{
downscale_intensity is implementing a convergence based downscaling
mechanism. This means that the intensity under investigation is assumed to
converge to a common value for all countries in a region at some point in
the future. Thus, the differences
}
\note{
The first year for which data is available is automatically taken as
the base year for the calculation of convergence. Thus, the data that is fed
in should always start at the intended base year.
}
\examples{

\dontrun{d1 <- downscale_intensity(gdpRegion,gdpCountry[,2010,],popCountry,mapping)}

\dontrun{d2 <- downscale_intensity(energyRegion,energyCountry[,2010,],gdpCountry,mapping)}

}
\seealso{
\code{\link{speed_aggregate}}
}
\author{
Roman Julius Hennig
}
