% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_filter.R
\name{gaussian_filter}
\alias{gaussian_filter}
\title{gaussian_filter}
\usage{
gaussian_filter(x,sigma=2,matrix_size=NULL)
}
\arguments{
\item{x}{The object to be smeared.}

\item{sigma}{The width of the gaussian filter in degree.)}

\item{matrix_size}{The size of the filter matrix in degree. All cells within
the range of the matrix size are taken into account for the smearing. For
NULL, the matrix size is 3 times sigma.}
}
\value{
The smeared MAgPIE object.
}
\description{
Smears cellular magpie objects with a gaussian filter.
}
\details{
The calculation is based on the \code{\link[raster]{focal}} function in the
\code{raster} library.
}
\examples{

 \dontrun{a <- gaussian_filter(x=land,sigma=3,matrix_size=4)}

}
\seealso{
\code{\link[raster]{focal}},\code{\link[raster]{focalWeight}},\code{\link{interpolate}}
}
\author{
Markus Bonsch
}
