% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrassProdDistr.R
\name{GrassProdDisagg}
\alias{GrassProdDisagg}
\title{Disaggregate managed pastures areas}
\usage{
GrassProdDisagg(grass_prod_lr,land_hr, weight_factor, map_file, lpjml_yields)
}
\arguments{
\item{grass_prod_lr}{grassland production in low resolution to be disaggregated}

\item{land_hr}{Magpie object with celular cost for each pasture management type}

\item{weight_factor}{Magpie object with celular yields for each management type}

\item{map_file}{cluster mappint to cells and regions}

\item{lpjml_yields}{Lpjml yeilds used for magpie optimization}
}
\value{
Mapie object with the total pasture areas optimized for fulfulling pasture demand in the available areas. An extra dimention is added to
capture cells that are infeasible.
}
\description{
This function calculates pasture management patters based on linear programing accounting for
costs of management, area available and production demands.
}
\examples{
\dontrun{
GrassProdDisagg(grass_prod_lr,land_hr, weight_factor, map_file, lpjml_yields)
}

}
\author{
Marcos Alves
}
