% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luscale-package.R
\docType{package}
\name{luscale-package}
\alias{luscale}
\alias{luscale-package}
\title{luscale: PIK Landuse Group Data Scaling Tools}
\description{
A collection of tools which allow to aggregate und disaggregate data in various ways.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/luscale}
  \item \doi{10.5281/zenodo.1158584}
  \item Report bugs at \url{https://github.com/pik-piam/luscale/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Benjamin Leon Bodirsky \email{bodirsky@pik-potsdam.de}
  \item Markus Bonsch
  \item Patrick, von Jeetze
  \item Ulrich Kreidenweiss
  \item Roman Julius Hennig
  \item Florian Humpenoeder \email{humpenoeder@pik-potsdam.de}
}

}
