% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag_hierarchical.R
\name{mag_hierarchical}
\alias{mag_hierarchical}
\title{MAgPIE Hierarchical clustering}
\usage{
mag_hierarchical(cdata, ncluster, ifolder, mode = "h", weight = NULL)
}
\arguments{
\item{cdata}{a cluster data file as produced by cluster_base}

\item{ncluster}{The desired total number of clusters.}

\item{ifolder}{input folder where the MAgPIE input files are located}

\item{mode}{Clustering type. At the moment you can choose between complete
linkage clustering (h), single linkage clustering (s) and Ward clustering
(w).}

\item{weight}{named vector with weighting factors for each region for the cluster distribution 
,e.g. weight=c(AFR=3,EUR=0.5). weight > 1 will grant more cluster to a region and
weight < 1 less cluster than by default.}
}
\value{
A spam relation matrix
}
\description{
Performs MAgPIE hierarchical clustering and calculates corresponding spam
relation matrix
}
\details{
As the creation of a clustering tree is very time consuming the function
checks first in the input folder if the corresponding data already exists
and if not it stores the tree information in the input folder for later use
in the next execution of this function.
}
\seealso{
\code{\link{cluster_per_region}}, \code{\link{mag_kmeans}},
\code{\link{clusterspam}}
}
\author{
Jan Philipp Dietrich
}
