% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag_kmeans.R
\name{mag_kmeans}
\alias{mag_kmeans}
\title{MAgPIE Kmeans clustering}
\usage{
mag_kmeans(cdata, ncluster = NULL, weight = NULL, cpr = NULL, seed = 42)
}
\arguments{
\item{cdata}{a cluster data file as produced by cluster_base}

\item{ncluster}{The desired total number of clusters.}

\item{weight}{named vector with weighting factors for each region for the cluster distribution 
,e.g. weight=c(AFR=3,EUR=0.5). weight > 1 will grant more cluster to a region and
weight < 1 less cluster than by default.}

\item{cpr}{cells-per-region information as returned by cluster_per_region. Weight and ncluster are
ignored in case that cpr is provided!}

\item{seed}{a single value, interpreted as an integer, or NULL, to define seed for random calculations}
}
\value{
A spam relation matrix
}
\description{
Performs MAgPIE kmeans clustering and calculates corresponding spam relation
matrix
}
\seealso{
\code{\link{cluster_per_region}}, \code{\link{mag_hierarchical}},
\code{\link{clusterspam}}
}
\author{
Jan Philipp Dietrich
}
