% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpieSums.R
\name{magpieSums}
\alias{magpieSums}
\title{Different summations of MAgPIE-objects}
\usage{
magpieSums(x, level="country", year="y1995")
}
\arguments{
\item{x}{A 0.5 degree MAgPIE-object}

\item{level}{Summation level. Currently available: "country", "region" and
"world"}

\item{year}{The year for which data should be summed up}
}
\value{
A matrix whith aggregation units as rows and data sums in columns
}
\description{
Returns the sum over all cells belonging to a specific aggregation unit
(countries,regions,world) of a 0.5 degree MAgPIE data set.
}
\examples{

  #magpieSums(x)

}
\seealso{
\code{\link{superAggregate}}
}
\author{
Jan Philipp Dietrich
}
