% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PastrMngtShares.R
\name{PastrMngtShares}
\alias{PastrMngtShares}
\title{Disaggregate managed pastures areas}
\usage{
PastrMngtShares(cost, yields, avl_area, production)
}
\arguments{
\item{cost}{Magpie object with cellular cost for each pasture management type}

\item{yields}{Magpie object with cellular yields for each management type}

\item{avl_area}{Magpie object with cellular data on the total managed pasture areas available}

\item{production}{Magpie object with cellular data on the total pasture demand}
}
\value{
Mapie object with the total pasture areas optimized to fulfill pasture demand in the available areas. An extra dimension is added to
capture cells that are unfeasible.
}
\description{
This function calculates pasture management patters based on linear programing accounting for
costs of management, area available and production demands.
}
\examples{
\dontrun{
PastrMngtShares(cost, yields, avl_area, production)
}

}
\author{
Marcos Alves
}
