% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spam.R
\name{read.spam}
\alias{read.spam}
\title{Read spam files}
\usage{
read.spam(file_name, file_type=NULL)
}
\arguments{
\item{file_name}{name of the spam-file}

\item{file_type}{File type, either "spam" or "sz" (for "spam-zipped"). File
ending is used if file_type is not mentioned}
}
\value{
\item{input_spam}{spam-object containing the transformation matrix
in spam-format}
}
\description{
Read transformation matrix from file. Data has the following structure:
\enumerate{ \item one 4byte Integer which contains the number of values
saved (n) \item n*2 matrix (integer) which contains the indices of the
spam-object (extracted with triplet(spam_object)) \item vector with n
elements (integer) containing the values of the spam-object \item nrows,
ncols (integer) }
}
\note{
Wildcards are supported in \code{file_name}
}
\examples{

\dontrun{
require(spam)
a <- matrix(c(0,1,1,0,0,1),3,2)
b <- as.spam(a)

print(b)
#      [,1] [,2]
# [1,]    0    0
# [2,]    1    0
# [3,]    1    1
# Class 'spam'

write.spam(b,"test.spam")

c <- read.spam("test.spam")

print(c)
#      [,1] [,2]
# [1,]    0    0
# [2,]    1    0
# [3,]    1    1
# Class 'spam'
file.remove("test.spam")

}
}
\seealso{
\code{\link{write.spam}}
}
\author{
Jan Philipp Dietrich
}
