% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_dimnames.r
\name{rename_dimnames}
\alias{rename_dimnames}
\title{rename_dimnames}
\usage{
rename_dimnames(data,dim=1,query=NULL,from=NULL,to=NULL)
}
\arguments{
\item{data}{Array}

\item{dim}{The dimension to be renamed.}

\item{query}{If NULL, query is automatically searched for. Otherwhise an
array, data.frame or the path of a csv with at least two columns. One column
has to have the name of "from", the other one the name of "to". Some queries
can be found in the svn-folder tools/queries.}

\item{from}{Only required if query is not NULL. Column of the query with
original dimnames of the incoming dataset}

\item{to}{Only required if query is not NULL. Column of the query with the
target dimnames of the outcoming dataset}
}
\value{
An array with different dimnames
}
\description{
Renames the dimnames of an array or MAgPIE object after a query.
}
\note{
translate.with.query has the same functionality, is more efficient,
yet more complicated to use.
}
\author{
Benjamin Bodirsky, Ulrich Kreidenweis
}
\keyword{array}
