% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_file.R
\name{reshape_file}
\alias{reshape_file}
\title{Reshape File}
\usage{
reshape_file(filename,spam=NULL,search_folders=NULL)
}
\arguments{
\item{filename}{file name of a file that should be disaggregated or an R
object to be disaggregated.}

\item{spam}{a vector of spam files and the corresponding needle. For more
details please check the sourcecode of this function with the predefined
spam vector. In case of an R object to be disagregated, no needle is
required.}

\item{search_folders}{Folders in which the script should search for spam
files. IMPORTANT: the paths are meant relative to the location where the
file given with fielname is stored! Not used if an R object is to be
disaggregated.}
}
\value{
NULL, if no disaggregation rule is found, otherwise the
disaggregated MAgPIE object
}
\description{
Disaggregates a cellular MAgPIE output to 0.5 degree based on spam matrices.
}
\examples{

 \dontrun{a <- reshape_file("cell.land.m.csv")}
\dontrun{a <- read.magpie("cell.land.m.csv")}
\dontrun{b <- reshape_file(a,spam="0.5_to_h100_area_weighted_mean.spam")}

}
\seealso{
\code{\link{reshape_folder}}, \code{\link{speed_aggregate}}
}
\author{
Jan Philipp Dietrich,Markus Bonsch
}
