% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_folder.R
\name{reshape_folder}
\alias{reshape_folder}
\title{Reshape Folder}
\usage{
reshape_folder(folder=".")
}
\arguments{
\item{folder}{folder in which the files should be disaggregated}
}
\description{
Disaggregates all cellular MAgPIE output in a folder to 0.5 degree files
based on spam matrices.
}
\examples{

 \dontrun{reshape_folder()}

}
\seealso{
\code{\link{reshape_file}}, \code{\link{speed_aggregate}}
}
\author{
Jan Philipp Dietrich
}
