% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean.groupAggregate.r
\name{weighted_mean.groupAggregate}
\alias{weighted_mean.groupAggregate}
\title{weighted_mean.groupAggregate}
\usage{
weighted_mean.groupAggregate(data, weight, dim, na.rm=TRUE, ...)
}
\arguments{
\item{data}{A MAgPIE object or array}

\item{weight}{An object with the same dimensions and dimnames as data.}

\item{dim}{The dimension over which is aggregated. Can be 1,3 or higher.
This dimension will be replaced by the larger categories specified in the
query.}

\item{na.rm}{removes NAs in data and weight}

\item{...}{additional arguments that will be handed on to groupAggregate().}
}
\value{
Returns a magpie object. The name dimension "dim" is grouped
according to the query, and a weighted mean is calculated for each
subset(group) of "dim".
}
\description{
Function which aggregates subsets of an array via an weighted mean function.
}
\examples{

a<-new.magpie(cells_and_regions = c("ZAF.1","ZAR.2","VEN.3"), 
              years = c("y1995","y2005"), 
              names = c("Wheat","Barley","Sugar_cane","Sugar_beet"), 
              fill=1:30)
#> a
#An object of class "magpie"
#, , Wheat
#
#      y1995 y2005
#ZAF.1     1     4
#ZAR.2     2     5
#VEN.3     3     6
#
#, , Barley
#
#      y1995 y2005
#ZAF.1     7    10
#ZAR.2     8    11
#VEN.3     9    12
#
#, , Sugar_cane
#
#      y1995 y2005
#ZAF.1    13    16
#ZAR.2    14    17
#VEN.3    15    18
#
#, , Sugar_beet
#
#      y1995 y2005
#ZAF.1    19    22
#ZAR.2    20    23
#VEN.3    21    24

b<-weighted_mean.groupAggregate(a,weight=a,dim=3)

#> b
#An object of class "magpie"
#, , tece
#
#      y1995     y2005
#ZAF.1  6.25  8.285714
#ZAR.2  6.80  9.125000
#VEN.3  7.50 10.000000
#
#, , sugr_cane
#
#      y1995 y2005
#ZAF.1    13    16
#ZAR.2    14    17
#VEN.3    15    18
#
#, , sugr_beet
# 
#      y1995 y2005
#ZAF.1    19    22
#ZAR.2    20    23
#VEN.3    21    24



}
\seealso{
\code{\link{groupAggregate}},\code{\link{colSums}},
\code{\link{magpieSums}}, \code{\link{superAggregate}}
}
\author{
Benjamin Bodirsky
}
