% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.spam.R
\name{write.spam}
\alias{write.spam}
\title{Write file in spam format}
\usage{
write.spam(input_spam, file_name, file_type=NULL)
}
\arguments{
\item{input_spam}{Spam-object, that should be written to file}

\item{file_name}{file name the object should be written to}

\item{file_type}{File type, either "spam" or "sz" (for "spam-zipped"). File
ending is used if file_type is not mentioned}
}
\description{
Write transformation matrix into file. Data has the following structure:
\enumerate{ \item one 4byte Integer which contains the number of values
saved (n) \item n*2 matrix (integer) which contains the indices of the
spam-object (extracted with triplet(spam\_object)) \item vector with n
elements (integer) containing the values of the spam-object \item nrows,
ncols (integer) }
}
\note{
Wildcards are supported in \code{file_name}
}
\examples{

require(spam)
a <- matrix(c(0,1,1,0,0,1),3,2)
b <- as.spam(a)

print(b)
#      [,1] [,2]
# [1,]    0    0
# [2,]    1    0
# [3,]    1    1
# Class 'spam'

write.spam(b,"test.spam")

c <- read.spam("test.spam")

print(c)
#      [,1] [,2]
# [1,]    0    0
# [2,]    1    0
# [3,]    1    1
# Class 'spam'
file.remove("test.spam")

}
\seealso{
\code{\link{read.spam}}
}
\author{
Jan Philipp Dietrich
}
