% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PastrMngtShares.R
\name{PastrMngtShares}
\alias{PastrMngtShares}
\title{Disaggregate managed pastures areas}
\usage{
PastrMngtShares(cost, yields, avl_area, production, eps)
}
\arguments{
\item{cost}{Magpie object with cellular cost for each pasture management type}

\item{yields}{Magpie object with cellular yields for each management type}

\item{avl_area}{Magpie object with cellular data on the total managed pasture areas available}

\item{production}{Magpie object with cellular data on the total pasture demand}

\item{eps}{epsilon Percent relaxation for area and production}
}
\value{
Mapie object with the total pasture areas optimized to fulfill pasture demand in the available areas. An extra dimension is added to
capture cells that are unfeasible.
}
\description{
This function calculates pasture management patters based on linear programing accounting for
costs of management, area available and production demands.
}
\details{
Data requirement
1. Costs for each level of pasture management
2. maximum production of each cell (the same as the maximum yields)
3. available pasture area in each cell

 Equations
 X[i] * C[i] = cost
 sum(X[i]) >= avl_area * (1 - eps)
 sum(X[i]) <= avl_area * (1 + eps)
 X[i] * Y[i] >= production * (1 - eps)
 X[i] * Y[i] <= production * (1 + eps)

Where: 
X : Pasture management areas
C : cost of production per type of management
Y : Pasture yields per type of managment
}
\examples{
\dontrun{
PastrMngtShares(cost, yields, avl_area, production, eps)
}

}
\author{
Marcos Alves
}
